backend/requirements.txt
anyio==4.8.0
asgiref==3.8.1
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==2.0.12
colorama==0.4.6
cryptography==44.0.3
defusedxml==0.7.1
Django==4.2.21
django-bootstrap5==22.2
django-cors-headers==4.7.0
django-filter==25.1
django_debug_toolbar==3.8.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
psycopg2-binary==2.9.9
exceptiongroup==1.2.2
Faker==12.0.1
flake8==5.0.4
flake8-docstrings==1.6.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mixer==7.2.2
oauthlib==3.2.2
packaging==23.0
pep8-naming==0.13.3
Pillow==9.3.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pycparser==2.22
pydocstyle==6.3.0
pyflakes==2.5.0
PyJWT==2.9.0
pyTelegramBotAPI==4.14.1
pytest==7.1.3
pytest-django==4.5.2
pytest-lazy-fixture==0.6.3
pytest-pythonpath==0.7.3
pytest-timeout==2.1.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==21.10
python3-openid==3.2.0
pytils==0.4.1
pytz==2022.7
requests==2.26.0
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
soupsieve==2.6
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
yapf==0.32.0
